fix rpg map types
