+++ packages/sample/src/server.ts
@@ -46,4 +46,4 @@ process.on('uncaughtException', function(error){
 
 process.on('unhandledRejection', function(reason: any){
     console.log(pe.render(reason)) 
-})
+})
