+++ packages/server/src/index.ts
@@ -13,9 +13,8 @@ import {
 import { RpgServer, RpgPlayerHooks, RpgServerEngineHooks } from './RpgServer'
 import { EventData } from './decorators/event'
 import { MapData } from './decorators/map'
+import { RpgPlayer, RpgEvent, EventMode } from './Player/Player'
 import { RpgMap } from './Game/Map'
-import { RpgEvent, EventMode } from './Event'
-import { RpgPlayer } from './Player/Player'
 import { Query } from './Query'
 import Monitor from './Monitor'
 import * as Presets from './presets'
